+++ bindings/python/tests-py/test_jsonschema.py
@@ -122,7 +122,7 @@ def test_error_message():
             str(exc)
             == """null is not of type "integer"
 
-Failed validating "type" in schema["properties"]["foo"]["type"]
+Failed validating "type" in schema["properties"]["foo"]
 
 On instance["foo"]:
     null"""